+++ Aoba/src/Core/TaskRunner/Timeout.cpp
@@ -6,7 +6,7 @@
 namespace s3d::aoba {
     std::atomic<size_t> Timeout::m_Id = 1;
 
-    Timeout::Timeout(const std::function<void()>& func, double ms, bool threading) :
+    Timeout::Timeout(const std::function<void()>& func, uint64_t ms, bool threading) :
         m_id(m_Id), m_ms(ms), m_threading(threading), m_func(func), m_startedTime(std::chrono::system_clock::now()) {
         m_Id++;
     }
@@ -14,7 +14,7 @@ namespace s3d::aoba {
     void Timeout::update() {
         if (isPending()) {
             const auto currentTime = std::chrono::system_clock::now();
-            const auto elapsedTime = static_cast<double>(
+            const auto elapsedTime = static_cast<uint64_t>(
                 std::chrono::duration_cast<std::chrono::milliseconds>(currentTime - m_startedTime).count());
             if (elapsedTime > m_ms) {
                 m_started = true;
